Initialize simple rectilinear grid

+++ staging/umwm_parallel.f90
@@ -1,7 +1,5 @@
 module umwm_parallel
 
-  use iso_fortran_env, only: int32, real32
-
   implicit none
 
   private
@@ -15,10 +13,10 @@ contains
 
   pure function denominators(n)
     !! Returns all common denominators of n.
-    integer(int32), intent(in) :: n
-    integer(int32), allocatable :: denominators(:)
-    integer(int32) :: i
-    denominators = [integer(int32) ::]
+    integer, intent(in) :: n
+    integer, allocatable :: denominators(:)
+    integer :: i
+    denominators = [integer ::]
     do i = 1, n
       if (mod(n, i) == 0) denominators = [denominators, i]
     end do
@@ -37,19 +35,19 @@ contains
     !!   * num_tiles(5) = [5, 1]
     !!   * num_tiles(6) = [3, 2]
     !!
-    integer(int32), intent(in) :: n
-    integer(int32) :: num_tiles(2)
-    integer(int32), allocatable :: denoms(:)
-    integer(int32), allocatable :: dim1(:), dim2(:)
-    integer(int32) :: i, j, n1, n2
+    integer, intent(in) :: n
+    integer :: num_tiles(2)
+    integer, allocatable :: denoms(:)
+    integer, allocatable :: dim1(:), dim2(:)
+    integer :: i, j, n1, n2
 
     ! find all common denominators of the total number of images
     denoms = denominators(n)
 
     ! Find all combinations of common denominators
     ! whose product equals the total number of images.
-    dim1 = [integer(int32) ::]
-    dim2 = [integer(int32) ::]
+    dim1 = [integer ::]
+    dim2 = [integer ::]
     do j = 1, size(denoms)
       do i = 1, size(denoms)
         if (denoms(i) * denoms(j) == n) then
@@ -74,9 +72,9 @@ contains
   pure function tile_indices_1d(dims, i, n) result(indices)
     !! Given input global array size, return start and end index
     !! of a parallel 1-d tile that correspond to this image.
-    integer(int32), intent(in) :: dims, i, n
-    integer(int32) :: indices(2)
-    integer(int32) :: offset, tile_size
+    integer, intent(in) :: dims, i, n
+    integer :: indices(2)
+    integer :: offset, tile_size
 
     tile_size = dims / n
 
@@ -98,9 +96,9 @@ contains
     !! Given an input x- and y- dimensions of the total computational domain
     !! [im, jm], returns an array of start and end indices in x- and y-,
     !! [is, ie, js, je].
-    integer(int32), intent(in) :: dims(2)
-    integer(int32) :: indices(4)
-    integer(int32) :: tiles(2), tiles_ij(2)
+    integer, intent(in) :: dims(2)
+    integer :: indices(4)
+    integer :: tiles(2), tiles_ij(2)
     tiles = num_tiles(num_images())
     tiles_ij = tile_n2ij(this_image())
     indices(1:2) = tile_indices_1d(dims(1), tiles_ij(1), tiles(1))
@@ -110,8 +108,8 @@ contains
 
   pure function tile_neighbors_1d() result(neighbors)
     !! Returns the image indices corresponding to left and right neighbor tiles.
-    integer(int32) :: neighbors(2)
-    integer(int32) :: left, right
+    integer :: neighbors(2)
+    integer :: left, right
     if (num_images() > 1) then
       left = this_image() - 1
       right = this_image() + 1
@@ -144,8 +142,8 @@ contains
     !!   * tile_n2ij(4) = [1, 2]
     !!   * tile_n2ij(6) = [3, 2]
     !!
-    integer(int32), intent(in) :: n
-    integer(int32) :: ij(2), i, j, tiles(2)
+    integer, intent(in) :: n
+    integer :: ij(2), i, j, tiles(2)
     if (n == 0) then
       ij = 0
     else
@@ -173,8 +171,8 @@ contains
     !!   * tile_ij2n([1, 2]) = 4
     !!   * tile_ij2n([3, 2]) = 6
     !!
-    integer(int32), intent(in) :: ij(2)
-    integer(int32) :: n, tiles(2)
+    integer, intent(in) :: ij(2)
+    integer :: n, tiles(2)
     if (any(ij == 0)) then
       n = 0
     else
@@ -187,10 +185,10 @@ contains
   pure function tile_neighbors_2d(periodic) result(neighbors)
     !! Returns the neighbor image indices given.
     logical, intent(in) :: periodic
-    integer(int32) :: neighbors(4)
-    integer(int32) :: tiles(2), tiles_ij(2), itile, jtile
-    integer(int32) :: left, right, down, up
-    integer(int32) :: ij_left(2), ij_right(2), ij_down(2), ij_up(2)
+    integer :: neighbors(4)
+    integer :: tiles(2), tiles_ij(2), itile, jtile
+    integer :: left, right, down, up
+    integer :: ij_left(2), ij_right(2), ij_down(2), ij_up(2)
 
     tiles = num_tiles(num_images())
     tiles_ij = tile_n2ij(this_image())
